requirements.txt
openai
chromadb
python-dotenv
pydantic
deepface
fastapi 
uvicorn
